Initial commit
For classes which `include Safety`, classes may create "safe" attributes
by calling `attr_accessor_` (or `reader` / `writer`), followed by an
optional type restriction by type constant (ie. `String`, `Fixnum`,
etc.) or a variable type declaration (`var`). A type safe `accessor`
restricted to `String`s would be `attr_accessor_String`. The first
parameter to this method is a symbol which is the name of the instance
variable / attribute (ie `:user_name`, `:product_count`, etc.). This
symbol becomes the name of the instance variable as well, but with the
`@` character appended (ie. `@user_name`, `@product_count`). Type
restricted attributes also guard against `nil` values (as `nil` is of
the type `NilClass`).

Either a default value or a lazy instantiation block may be passed to
the attribute creation method as well. This will set the instance
variable to either the default value or the returned value of the lazy
instantiation block (preference is given to lazy instantiation). The
default value is passed as the second (optional) parameter, and the
block can be passed as a normal `do` block or with the block
reference syntax `&`.